On branch ver-a1.0  Changes to be committed: 	modified:   _layouts/default.html 	renamed:    pages/school_advai.md -> pages/school/school_advai.md 	renamed:    pages/school_ml.md -> pages/school/school_ml.md 	renamed:    pages/school_nlp.md -> pages/school/school_nlp.md 	renamed:    pages/school_se.md -> pages/school/school_se.md
New branch, ver-a1.0, begins updates of the navigation pane and website in general in relation to the plan for quickchat commercial version coverage, along with the free version information, tutorials on various topics including the programs, programming, ai, etc...  discussions will potentially be done on github through the discussions feature, which might be referred to as issues, researching this topic.  information will be on the site about this discussion area.